pyproject.toml
[project]
name = "blender-cabinet-nodes"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "fake-bpy-module-latest>=20251003",
]
